requirements.txt
polygon-api-client
pandas
matplotlib
